requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
lightgbm==4.5.0
scikit-learn==1.5.2
imbalanced-learn==0.12.4
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2
httpx==0.27.2
